move LogOut from Authentication to Header
