+++ src/components/Pages/AuthenticationPage/validator.js
